@slaren 2024-06-10
Looks very good. Should MMQ be the default again?

| GPU         | Model     |   Microbatch size | Test   |   t/s master cuBLAS |   t/s cuda-ptx-mma-12 |   Speedup |
|:------------|:----------|------------------:|:-------|-------------:|----------------------:|----------:|
| RTX 3090 Ti | 7B Q4_K_M |                16 | pp1024 |       763.74 |               1106.51 |      1.45 |
| RTX 3090 Ti | 7B Q4_K_M |                32 | pp1024 |      1098.05 |               1824.08 |      1.66 |
| RTX 3090 Ti | 7B Q4_K_M |                64 | pp1024 |      1521.28 |               2423.70 |      1.59 |
| RTX 3090 Ti | 7B Q4_K_M |               128 | pp1024 |      2437.95 |               3051.37 |      1.25 |
| RTX 3090 Ti | 7B Q4_K_M |               256 | pp1024 |      3651.28 |               3724.66 |      1.02 |
| RTX 3090 Ti | 7B Q4_K_M |               512 | pp1024 |      4290.68 |               4208.15 |      0.98 |
| RTX 3090 Ti | 7B Q4_K_M |              1024 | pp1024 |      4730.21 |               4150.73 |      0.88 |
| RTX 3090 Ti | 7B Q5_K_M |                16 | pp1024 |       704.44 |                986.14 |      1.40 |
| RTX 3090 Ti | 7B Q5_K_M |                32 | pp1024 |       998.54 |               1557.86 |      1.56 |
| RTX 3090 Ti | 7B Q5_K_M |                64 | pp1024 |      1345.34 |               2128.63 |      1.58 |
| RTX 3090 Ti | 7B Q5_K_M |               128 | pp1024 |      2404.66 |               2833.24 |      1.18 |
| RTX 3090 Ti | 7B Q5_K_M |               256 | pp1024 |      3574.42 |               3543.32 |      0.99 |
| RTX 3090 Ti | 7B Q5_K_M |               512 | pp1024 |      4259.87 |               3906.15 |      0.92 |
| RTX 3090 Ti | 7B Q5_K_M |              1024 | pp1024 |      4695.34 |               3871.94 |      0.82 |
| RTX 3090 Ti | 7B Q6_K   |                16 | pp1024 |       629.23 |                910.99 |      1.45 |
| RTX 3090 Ti | 7B Q6_K   |                32 | pp1024 |       985.37 |               1523.31 |      1.55 |
| RTX 3090 Ti | 7B Q6_K   |                64 | pp1024 |      1355.43 |               2203.05 |      1.63 |
| RTX 3090 Ti | 7B Q6_K   |               128 | pp1024 |      2385.65 |               3010.48 |      1.26 |
| RTX 3090 Ti | 7B Q6_K   |               256 | pp1024 |      3566.80 |               3654.09 |      1.02 |
| RTX 3090 Ti | 7B Q6_K   |               512 | pp1024 |      4215.07 |               3999.10 |      0.95 |
| RTX 3090 Ti | 7B Q6_K   |              1024 | pp1024 |      4617.20 |               3935.87 |      0.85 |