@Green-Sky 2024-06-10
@JohannesGaessler both your comparison tables are vs master mmq.