@JohannesGaessler 2024-06-10
>Looks very good. Should MMQ be the default again?

Give me a bit more time to implement q2_K and q3_K and to optimize performance (particularly asynchronous data loading). Then I think MMQ will be universally faster. Also, in case you're not aware, int8 tensor cores are only available with Ampere (rather than Volta). So for V100s FP16 cuBLAS should still be the fastest option.

>@JohannesGaessler both your comparison tables are vs master mmq.

Only for batch sizes 16, 32, and 64; I compared vs. with/without `LLMA_CUDA_FORCE_MMQ` and currently even without that option MMQ is used for batch sizes <= 64 on master.