@ggerganov 2024-06-11
For BS >= 512 F16 cuBLAS is still faster even with tensor cores, is that correct?