@JohannesGaessler 2024-06-11
It depends on the quantization format and hardware; q4_0 on an RTX 4090 seems to be the best case scenario where it seems to already be faster even for large batch sizes.

FP16 cuBLAS:

| model                          |       size |     params | backend    | ngl |   n_ubatch |         fa |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------: | ------------: | ---------------: |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | CUDA       |  99 |        512 |          1 |        pp4096 |   7430.94 ± 0.00 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | CUDA       |  99 |       1024 |          1 |        pp4096 |   8655.73 ± 0.00 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | CUDA       |  99 |       2048 |          1 |        pp4096 |   8798.88 ± 0.00 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | CUDA       |  99 |       4096 |          1 |        pp4096 |   8790.74 ± 0.00 |

int8 tensor core MMQ:

| model                          |       size |     params | backend    | ngl |   n_ubatch |         fa |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------: | ---------: | ------------: | ---------------: |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | CUDA       |  99 |        512 |          1 |        pp4096 |   9719.23 ± 0.00 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | CUDA       |  99 |       1024 |          1 |        pp4096 |   9692.96 ± 0.00 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | CUDA       |  99 |       2048 |          1 |        pp4096 |   9094.65 ± 0.00 |
| llama 8B Q4_0                  |   4.33 GiB |     8.03 B | CUDA       |  99 |       4096 |          1 |        pp4096 |   9115.52 ± 0.00 |