@sorasoras 2024-06-11
@JohannesGaessler Are you plan to do the same for IQ quants？ It would be nice to run Int8 on my P40 instead of FP32. IQ quant has been very slow  on that card