@JohannesGaessler 2024-06-11
I will prioritize the quantization formats that already have MMQ implementations (legacy, k-quants) but long-term I plan to also implement kernels for the other quantization formats.