@Dampfinchen 2024-06-17
Benchmark on Turing (RTX 2060, FA, batch size default, 4096 context, q4_k_s)

Cublas 

```
llama_print_timings:        load time =    3416.67 ms
llama_print_timings:      sample time =      21.96 ms /   180 runs   (    0.12 ms per token,  8195.23 tokens per second)
llama_print_timings: prompt eval time =    4760.33 ms /  3671 tokens (    1.30 ms per token,   771.17 tokens per second)
llama_print_timings:        eval time =    4922.92 ms /   179 runs   (   27.50 ms per token,    36.36 tokens per second)
llama_print_timings:       total time =    9861.65 ms /  3850 tokens
```

Force_MMQ

`llama_print_timings:        load time =    3430.94 ms
llama_print_timings:      sample time =      22.20 ms /   180 runs   (    0.12 ms per token,  8109.57 tokens per second)
llama_print_timings: prompt eval time =    7682.43 ms /  3671 tokens (    2.09 ms per token,   477.84 tokens per second)
llama_print_timings:        eval time =    4949.56 ms /   179 runs   (   27.65 ms per token,    36.16 tokens per second)
llama_print_timings:       total time =   12801.45 ms /  3850 tokens`

Looks like MMQ's prompt processing is still quite a bit slower. I've tested this with the most up to date build at the time of this writing.