@JohannesGaessler 2024-06-17
Did you use make or cmake to build the project? As of right now cmake compiles for the wrong CUDA architectures so the int8 tensor cores aren't going to actually be used.