@JohannesGaessler 2024-06-17
Lines 426 and 428 in `CMakeLists.txt`, replace `70` with `75`.