@Dampfinchen 2024-06-17
> Lines 426 and 428 in `CMakeLists.txt`, replace `70` with `75`.

```
llama_print_timings:        load time =    3430.78 ms
llama_print_timings:      sample time =      22.07 ms /   180 runs   (    0.12 ms per token,  8155.13 tokens per second)
llama_print_timings: prompt eval time =    3996.73 ms /  3671 tokens (    1.09 ms per token,   918.50 tokens per second)
llama_print_timings:        eval time =    4912.27 ms /   179 runs   (   27.44 ms per token,    36.44 tokens per second)
llama_print_timings:       total time =    9080.10 ms /  3850 tokens
```

Yep, that was it. It's faster than cublas now, wow. Great result!!